Improve things a bit
